+++ src/Sentry.AspNet/SentryAspNetOptionsExtensions.cs
@@ -23,6 +23,9 @@ namespace Sentry.AspNet
 
             var eventProcessor = new SystemWebRequestEventProcessor(payloadExtractor, options);
 
+            // Ignore options.IsGlobalModeEnable, we always want to use HttpContext as backing store here
+            options.ScopeStackContainer ??= new HttpContextScopeStackContainer();
+
             options.DiagnosticLogger ??= new TraceDiagnosticLogger(options.DiagnosticLevel);
             options.Release ??= SystemWebVersionLocator.Resolve(options, HttpContext.Current);
             options.AddEventProcessor(eventProcessor);