+++ src/Sentry/ITransactionContext.cs
@@ -9,5 +9,10 @@
         /// Transaction name.
         /// </summary>
         string Name { get; }
+
+        /// <summary>
+        /// Whether the parent transaction of this transaction has been sampled.
+        /// </summary>
+        bool? IsParentSampled { get; }        
     }
 }