+++ src/Sentry/Transaction.cs
@@ -49,6 +49,9 @@ namespace Sentry
         public string Name { get; private set; }
 
         /// <inheritdoc />
+        public bool? IsParentSampled { get; set; }
+
+        /// <inheritdoc />
         public string? Platform { get; set; } = Constants.Platform;
 
         /// <inheritdoc />
@@ -197,11 +200,15 @@ namespace Sentry
         /// Initializes an instance of <see cref="Transaction"/>.
         /// </summary>
         public Transaction(ITransaction tracer)
-            : this(tracer.Name, tracer.Operation)
+            : this(tracer.Name)
         {
+            // Contexts have to be set first because other fields use that
+            Contexts = tracer.Contexts;
+
             ParentSpanId = tracer.ParentSpanId;
             SpanId = tracer.SpanId;
             TraceId = tracer.TraceId;
+            Operation = tracer.Operation;
             Platform = tracer.Platform;
             Release = tracer.Release;
             StartTimestamp = tracer.StartTimestamp;
@@ -211,7 +218,6 @@ namespace Sentry
             IsSampled = tracer.IsSampled;
             Level = tracer.Level;
             Request = tracer.Request;
-            Contexts = tracer.Contexts;
             User = tracer.User;
             Environment = tracer.Environment;
             Sdk = tracer.Sdk;